API Reference

Storage

Optuna.RDBStorage — Type

RDBStorage(database_path::String, database_name::String)

Storage class for RDB backends.

Arguments

database_path::String: Path to the directory where the database is stored.

database_name::String: Name of the database file.

Optuna.get_all_study_names — Function

get_all_study_names(storage::BaseStorage)

returns all study names stored in the given storage.

Arguments

storage::BaseStorage: Storage backend to query.

Returns

Vector{String}: List of study names.

Artifacts

Optuna.FileSystemArtifactStore — Type

FileSystemArtifactStore(path::String)

Data structure for a file system based artifact store.

Arguments

path::String: Path to the directory where artifacts are stored.

Optuna.ArtifactMeta — Type

ArtifactMeta

Data structure containing metadata for an artifact.

Sampler

Optuna.RandomSampler — Type

RandomSampler(seed=nothing::Union{Nothing,Int})

An independent sampler that samples randomly

Arguments

seed::Union{Nothing,Int}=nothing: Seed for the random number generator.

Pruner

Optuna.MedianPruner — Type

MedianPruner(n_startup_trials::Int=5,
    n_warmup_steps::Int=0,
    interval_steps::Int=1;
    n_min_trials::Int=1)

Prune a trial if its best intermediate result is below the median of intermediate results of previous completed trials at the same step.

Arguments

n_startup_trials::Int=5: Number of initial trials before pruning starts.

n_warmup_steps::Int=0: Number of steps in a trial before pruning starts.

interval_steps::Int=1: Interval in number of steps when the pruner checks, offset by n_warmup_steps.

n_min_trials::Int=1: Minimum trials needed at a step to trigger pruning. If reported results are below n_min_trials, the trial continues regardless of performance.

Study

Optuna.Study — Type

This data structure represents an Optuna study and its corresponding artifact and data storage. A study is a collection of trials that share the same optimization objective.

Optuna.load_study — Function

load_study(
    storage::BaseStorage,
    study_name::String,
    artifact_store::BaseArtifactStore;
    sampler=nothing::Union{Nothing,BaseSampler},
    pruner=nothing::Union{Nothing,BasePruner},
)

Load an existing study with the given name, artifact store, storage, sampler and pruner.

Arguments

study_name::String: Name of the study.

artifact_store::BaseArtifactStore: Artifact store for the study. (see Artifacts)

storage::BaseStorage: Storage of the study. (see Storage)

Keyword Arguments

sampler::Union{Nothing,BaseSampler}=nothing: Sampler to use for the study. (see Sampler)

pruner::Union{Nothing,BasePruner}=nothing: Pruner to use for the study. (see Pruner)

Returns

Study: The loaded study. (see Study)

Optuna.delete_study — Function

delete_study(
    study_name::String,
    storage::BaseStorage,
)

Delete a study with the given name and storage backend.

Arguments

study_name::String: Name of the study.

storage::BaseStorage: Storage of the study. (see Storage)

Optuna.copy_study — Function

copy_study(
    from_study_name::String,
    from_storage::BaseStorage,
    to_storage::BaseStorage,
    to_study_name::String="",
)

Copy a study from one storage backend to another.

Arguments

from_study_name::String: Name of the study to copy.

from_storage::BaseStorage: Storage backend to copy the study from. (see Storage)

to_storage::BaseStorage: Storage backend to copy the study to. (see Storage)

to_study_name::String="": Name of the new study. If empty, the original study name is used.

Optuna.ask — Function

ask(study::Study)

Get a new trial from the study. This is an alternative way to create trials instead of using the optimize function. A trial created this way must be completed using the tell function.

Arguments

study::Study: The study to get the trial from. (see Study)

Returns

Trial: The new trial. (see Trial)

Optuna.tell — Function

tell(study::Study, trial::Trial, score::Union{Nothing,T,Vector{T}}=nothing; prune::Bool=false

) where {T<:AbstractFloat}

Tell the study about the result of a trial. This is the proper way to complete a trial created with the ask function.

Arguments

study::Study: The study to report the trial to. (see Study)

trial::Trial: The trial which was completed. (see Trial)

score::Union{Nothing,T,Vector{T}}=nothing: The score of the trial. If nothing, the trial is pruned.

prune::Bool=false: If true, the trial is pruned.

Optuna.best_trial — Function

best_trial(study::Study)

Get the best trial of the study.

Arguments

study::Study: The study to get the best trial from. (see Study)

Returns

trial: The best trial. (see Trial)

Optuna.best_params — Function

best_params(study::Study)

Get the best parameters of the study.

Arguments

study::Study: The study to get the best parameters from. (see Study)

Returns

parameters: The best parameters.

Optuna.best_value — Function

best_value(study::Study)

Get the best objective value of the study.

Arguments

study::Study: The study to get the best objective value from. (see Study)

Returns

value: The best objective value.

Optuna.upload_artifact — Function

upload_artifact(study::Study, trial::Trial, data::Dict)

Upload an artifact for a given trial in the study. The artifact is a .jld2 file containing the provided data.

Arguments

study::Study: The study to upload the artifact to. (see Study)

trial::Trial: The trial to associate the artifact with. (see Trial)

data::Dict: The data to be stored as an artifact.

Optuna.get_all_artifact_meta — Function

get_all_artifact_meta(study::Study)

Get all artifact metadata for all trials in the given study.

Arguments

study::Study: The study to get the artifact metadata from. (see Study)

Returns

Vector{ArtifactMeta}: List of artifact metadata of all artifacts in the study.

get_all_artifact_meta(study::Study, trial)

Get all artifact metadata for the trial in the given study.

Arguments

study::Study: The study to get the artifact metadata from. (see Study)

trial: The trial to get the artifact metadata from. (see Trial)

Returns

Vector{ArtifactMeta}: List of artifact metadata of the given trial.

Optuna.download_artifact — Function

download_artifact(study::Study, artifact_id::String, file_path::String)

Download an artifact from a given study identified by its artifactid and the filepath where it should be stored.

Arguments

study::Study: The study to download the artifact from. (see Study)

artifact_id::String: The ID of the artifact to download.

file_path::String: The path where the downloaded artifact should be stored.

Trial

Optuna.Trial — Type

Trial is a data structure wrapper for an Optuna trial.

Optuna.suggest_int — Function

suggest_int(trial::Trial, name::String, low::T, high::T) where {T<:Signed}

Suggest an integer value for the given parameter name within the specified range.

Arguments

trial::Trial: The trial to suggest the parameter for. (see Trial)

name::String: The name of the parameter to suggest.

low::T: The lower bound of the range (inclusive).

high::T: The upper bound of the range (inclusive).

Returns

T: Suggested integer value.

Optuna.suggest_float — Function

suggest_float(trial::Trial, name::String, low::T, high::T) where {T<:AbstractFloat}

Suggest a float value for the given parameter name within the specified range.

Arguments

trial::Trial: The trial to suggest the parameter for. (see Trial)

name::String: The name of the parameter to suggest.

low::T: The lower bound of the range (inclusive).

high::T: The upper bound of the range (inclusive).

Returns

T: Suggested float value.

Optuna.suggest_categorical — Function

suggest_categorical(trial::Trial, name::String, choices::Vector{T}) where {T<:Union{Bool,Int,AbstractFloat,String}}

Suggest a categorical value for the given parameter name from the specified choices.

Arguments

trial::Trial: The trial to suggest the parameter for. (see Trial)

name::String: The name of the parameter to suggest.

low::T: The lower bound of the range (inclusive).

high::T: The upper bound of the range (inclusive).

Returns

T: Suggested integer value.

Optuna.report — Function

report(trial::Trial, value::AbstractFloat, step::Int)

Report an intermediate value for the given trial at a specific step.

Arguments

trial::Trial: The trial to report the value for. (see Trial)

value::AbstractFloat: The intermediate value to report.

step::Int: The step at which the value is reported.

Optuna.should_prune — Function

should_prune(trial::Trial)

Check if the given trial should be pruned based on the pruner's decision.

Arguments

trial::Trial: The trial to check for pruning. (see Trial)

Returns

Bool: true if the trial should be pruned, false otherwise.

Optimization

Optuna.optimize — Function

optimize(study::Study, objective::Function, params::NamedTuple; n_trials=100::Int)

Optimize the objective function by choosing a suitable set of hyperparameter values from the given params.

Uses the sampler of the study which implements the task of value suggestion based on a specified distribution. For the available samplers see Sampler.

If the objective function returns nothing, the trial is pruned.

Arguments

study::Study: Study that should be optimized.

objective::Function: Function that takes a trial and returns a score.

params::NamedTuple: Named tuple of parameters to optimize.

Keyword Arguments

n_trials::Int: Number of trials to run.

Returns

Study: The optimized study. Study